% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_bayes_inference.R
\name{beta_binom}
\alias{beta_binom}
\title{Bayesian estimation for beta binomial distribution}
\usage{
beta_binom(n, y, a = 1, b = 1)
}
\arguments{
\item{n}{Total number of binomial draws}

\item{y}{The total number of successes}

\item{a}{Shape 1 parameter of beta distribution}

\item{b}{Shape 1 parameter of beta distribution}
}
\value{
The likelihood and posterior of the beta binmomial distribution
}
\description{
This function returns the bayesian posterior estimate for a beta binomial distribution
}
