% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_factorize_matrix.R
\name{check_inverse}
\alias{check_inverse}
\title{Can the matrix be inverted?}
\usage{
check_inverse(m)
}
\arguments{
\item{m}{A Matrix}
}
\value{
LOGICAL Indicating invertibility of matrix
}
\description{
This function checks if target matrix is invertible
}
