% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_factorize_matrix.R
\name{col_scale}
\alias{col_scale}
\title{Column normalization}
\usage{
col_scale(dat, select)
}
\arguments{
\item{dat}{A DataFrame}

\item{select}{A vector of selected column names for normalization}
}
\value{
A DataFrame just with selected columns standardized
}
\description{
This function standardizes only certain columns of an existing dataframe (specified by the user).
Standardization involves scaling columnwise.
}
