% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_tools.R
\name{color_hclust}
\alias{color_hclust}
\title{Hclust in colour}
\usage{
color_hclust(hclust, labels = hclust$labels, lab.col = rep(1,
  length(hclust$labels)), hang = 0.1, xlab = "", sub = "")
}
\arguments{
\item{hclust}{hclust object}

\item{labels}{a character vector of labels of the leaves of the tree}

\item{lab.col}{colour for the labels; NA=default device foreground colour}

\item{xlab}{title for x-axis}

\item{sub}{subtitle}
}
\value{
Function
}
\description{
Prints pretty hclust plots
}
