% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_factorize_matrix.R
\name{counts_to_cpm}
\alias{counts_to_cpm}
\title{Compute counts per million (CPM)}
\usage{
counts_to_cpm(expr_mat, spikes = NULL)
}
\arguments{
\item{expr_mat}{Input matrix with expression values for genes in the form (Genes x Samples)}

\item{spikes}{A vector of genes that were spike-ins in the original experiment}
}
\value{
A matrix with CPM expression values
}
\description{
Convert raw counts to counts per million (CPM) by dividing each column by its total then multiplying by 1,000,000.
Spike-ins should be excluded from the calculation of total expression in order to correct for total cell RNA content, therefore they should be specified to exclude them from downstream analysis
}
