% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_factorize_matrix.R
\name{counts_to_rle}
\alias{counts_to_rle}
\title{Size Factor normalization as used by DESeq for differential expression. Also called RLE (Relative Log Expression)}
\usage{
counts_to_rle(expr_mat, spikes = NULL)
}
\arguments{
\item{expr_mat}{Input matrix with expression values for genes in the form (Genes x Samples)}

\item{spikes}{A vector of genes that were spike-ins in the original experiment}
}
\value{
A matrix with SF normalized RLE values
}
\description{
First the geometric mean of each gene across all cells is calculated. The size factor for each cell is the median across genes of the ratio of the expression to the gene’s geometric mean.
Spike-ins should be excluded from the calculation of total expression in order to correct for total cell RNA content, therefore they should be specified to exclude them from downstream analysis
}
