% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_tools.R
\name{fancy_heatmap}
\alias{fancy_heatmap}
\title{Fancy Heatmap}
\usage{
fancy_heatmap(X, top, ...)
}
\arguments{
\item{X}{Gene x Sample Matrix}

\item{top}{Pick top genes with highest variability}

\item{...}{Additional arguments to pass to heatmap.2}
}
\value{
Function
}
\description{
Plots a heatmap of clstering within a Gene x Sample Matrix of top n genes.
}
