% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast_load_write.R
\name{fastload_mat}
\alias{fastload_mat}
\title{Load a File as Matrix}
\usage{
fastload_mat(infile)
}
\arguments{
\item{infile}{Path to the input file}
}
\value{
A matrix of the input file
}
\description{
This function loads a file as a matrix. It assumes that the first column
contains the rownames and the subsequent columns are the sample identifiers.
Any rows with duplicated row names will be dropped with the first one being kept.
}
