% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast_load_write.R
\name{fastwrite_tab}
\alias{fastwrite_tab}
\title{Write a New File as Tab Delimited}
\usage{
fastwrite_tab(outmat, outfile)
}
\arguments{
\item{outmat}{Path to the output object}
}
\description{
This function writes to a file. It assumes that the object to be written #' is a data.table object. It handles lists or matrix formats.
Column identifiers are retained.
}
