% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_attributes_functions.R
\name{fetch_gc_content}
\alias{fetch_gc_content}
\title{Fetch GC content from gene IDs}
\usage{
fetch_gc_content(select_genes)
}
\arguments{
\item{select_genes}{A character vector with a given set of interesting genes}
}
\value{
A vector of gene GC contents
}
\description{
A Genomic Ranges object is constructed from  UCSC Genome Browser, identified by the Ensembl hg19 assembly gene IDs.
The function then retrieves the sequences of individual genes identified by the corrdinates of the gene IDs from BSgenome hg19 assembly.
The alphabet frequencies are computed, and then the G+C content is reported as a fraction. This is returned per gene ID
}
