% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_attributes_functions.R
\name{fetch_gene_length}
\alias{fetch_gene_length}
\title{Fetch gene lengths from gene IDs}
\usage{
fetch_gene_length(select_genes)
}
\arguments{
\item{select_genes}{A character vector with a given set of interesting genes}
}
\value{
A vector of gene lengths
}
\description{
Using the Ensembl hg19 assembly, the function established a SQL connection with the Ensembl DB, and retrieves all the transcripts associated with each gene.
A longest transcript is determined by merging all the exons of individual transcripts. 
The concatenated length is computed and returned per gene ID
}
