% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesis_tests.R
\name{foo_tests}
\alias{foo_tests}
\title{Hypothesis tests}
\usage{
foo_tests(x, y, tests)
}
\arguments{
\item{x}{Matrix 1}

\item{y}{Matrix 2}

\item{tests}{A vector with one or more of the following values: 1) t (t-test), 2) f (f-test), 3) mean (difference of means), 4) var (difference of variance, 5) wilcox (wilcoxon test), 6) ks (KS test)}
}
\value{
A data frame with the indices resulting from ordering the vector of, one of 1) differences, 2) statistics, or 3) pvalues computed between matched columns of the two matrices, in decreasing order of magnitude.
}
\description{
Carries out hypotheses tests testing for significant differences in matched columns
of two matrices, and returns the most significant to the least significant result in order.
}
