% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_attributes_functions.R
\name{GO_analysis}
\alias{GO_analysis}
\title{Gene Ontology Analysis}
\usage{
GO_analysis(all_genes, select_genes)
}
\arguments{
\item{all_genes}{A character vector with all genes}

\item{select_genes}{A character vector with a given set of interesting genes}
}
\value{
A topGO GenTable object with fold enrichment, pvalue, adjusted p-value and FDR.
}
\description{
This function conducts GO analysis for a set of selevcted genes.
The background genes can also be changed.
}
