% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variant_analysis.R
\name{GWAS}
\alias{GWAS}
\title{GWAS}
\usage{
GWAS(genodata = genotypes, phenodata = phenotypes, family = gaussian,
  filename = NULL, append = FALSE, workers = getOption("mc.cores", 2L),
  flip = TRUE, select.snps = NULL, nSplits = 10)
}
\arguments{
\item{genodata}{Specify the entire genotype data object in SnpMatrix format}

\item{phenodata}{Data frame with a column of sample IDs, corresponding to the row names of genodata, and a columns for the continuous outcome variable. These columns must be named "id" and "phenotype".}

\item{family}{Specifies the link function for GLM. Default is guassian}

\item{filename}{Output filename with path}

\item{append}{LOGICAL whether to append output}

\item{workers}{Number of cores to use for parallel processing of function}

\item{flip}{In order to fit the model, genotype data is converted to numeric format using the as function from snpStats. The genotypes of each SNP are then coded as continuous, thereby taking on the value of 0, 1, and 2. To indicate the number of minor alleles instead, a flip.matrix procedure is included, which can be turned on or off using the flip argument.}

\item{select.snps}{Subset the analysis via a vector of SNP IDs}

\item{nSplits}{Number of SNP-wise splits that are made to the genotype data. The function runs the GWA analysis on these smaller subsets of the genotype data one at a time. After each subset has finished running the function will print a progress update onto the R console. By default this is set to 10}
}
\value{
Outputs results in file with columns: "SNP", "Estimate", "Std.Error", "t-value", "p-value"
}
\description{
Conducts GWAS (per SNP association) for given genotype-phenotype data.
}
