% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_bayes_inference.R
\name{logistic.model}
\alias{logistic.model}
\title{Implementing binomial GLM (logistic regression) from scratch.}
\usage{
logistic.model(X, Y)
}
\arguments{
\item{X}{A Covariate vector of values.}

\item{Y}{A Response vector of values.}
}
\value{
A vector containing the optimized MLE parameters of the fit as well as the AIC and BIC to enable downstream model selection
}
\description{
This function accepts a covariate X and response Y and uses optimization to estimate the maximum likelihood parameters of the binomial GLM fit.
General-purpose optimization to estimate parameters is based on Nelder–Mead algorithm
The Akaike and Bayesian Information Criterion is also output to enable downstream model selection
}
