% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_tools.R
\name{ma_plot}
\alias{ma_plot}
\title{MA plot}
\usage{
ma_plot(x, y, n = 10000, subset = NULL, xlab = NULL, ylab = NULL,
  curve.add = TRUE, curve.col = 2, curve.span = 1/2, curve.lwd = 2,
  curve.n = 2000, ...)
}
\arguments{
\item{x}{a numeric vector}

\item{y}{a numeric vector}

\item{n}{a numeric value to subsample data}

\item{subset}{index of the points to be plotted}

\item{xlab}{a title for the x axis}

\item{ylab}{a title for the y axis}

\item{curve.add}{if TRUE a smooth curve is fit to the data and displayed. The function loess is used to fit the curve.}

\item{curve.col}{a numeric value that determines the color of the smooth curve}

\item{curve.span}{is passed on to loess}

\item{curve.lwd}{the line width for the smooth curve}

\item{curve.n}{a numeric value that determines the sample size used to fit the curve. This makes fitting the curve faster with large datasets}

\item{...}{further arguments passed to plot}
}
\value{
Function
}
\description{
Takes two vectors x and y and plots 
M=y-x versus A=(x+y)/2. 
A smooth curve is added to show trends.
}
