% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_factorize_matrix.R
\name{min_max_scale}
\alias{min_max_scale}
\title{Minimum maximum scaling of vector}
\usage{
min_max_scale(x)
}
\arguments{
\item{x}{A Vector}
}
\value{
A Vector scaled by maximum value.
}
\description{
This function loads a file as a matrix. It assumes that the first column
contains the rownames and the subsequent columns are the sample
}
