% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_bayes_inference.R
\name{mix.beta}
\alias{mix.beta}
\title{Bayesian estimation for a mixture of betas.}
\usage{
mix.beta(x, p, a, b)
}
\arguments{
\item{x}{Number of observations per beta distribution}

\item{p}{Number of betas}

\item{a}{A vector of shape 1 parameters for all beta distributions}

\item{b}{A vector of shape 2 parameters for all beta distributions}
}
\value{
The combined posterior estimate of the mixtures
}
\description{
This function returns the bayesian posterior estimate for a mixture of beta distributions
}
