% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_bayes_inference.R
\name{mle_gaussian_dist}
\alias{mle_gaussian_dist}
\title{Gaussian MLE parameters and Model selection}
\usage{
mle_gaussian_dist(X, start_mu, start_sigma)
}
\arguments{
\item{X}{A Vector of values.}

\item{start_mu}{Initial value to begin optimization at for mean of gaussian distribution}

\item{start_sigma}{Initial value to begin optimization at for standard deviation of gaussian distribution}
}
\value{
A vector containing the optimized MLE mean and standard deviation as well as the log likelihood and AIC to enable downstream model selection
}
\description{
This function accepts a vector with all values of a given unknown distribution
It estimates the maximum likelihood parameters by fitting a guassian distribution over it.
General-purpose optimization to estimate parameters is based on Nelder–Mead algorithm
The Akaike Information Criterion is also output to enable downstream model selection
}
