% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_factorize_matrix.R
\name{plot_mds}
\alias{plot_mds}
\title{Multi-Dimensional Scaling 2D plot}
\usage{
plot_mds(X)
}
\arguments{
\item{X}{A Matrix}
}
\value{
A 2D MDS plot
}
\description{
This function accepts a matrix, computes the pairwise distance matrix, and projects the distances onto a lwer dimensional (2D space)
}
