% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_factorize_matrix.R
\name{plot_pca}
\alias{plot_pca}
\title{Principal Component Analysis 2D plot}
\usage{
plot_pca(X, ...)
}
\arguments{
\item{X}{A Gene Matrix}
}
\value{
A plot with the first two eigen vectors plotted on the X and Y axis
}
\description{
This function accepts a matrix, computes the principal components using a kernel transformation and plots the first 2 eigen vectors
}
