% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_factorize_matrix.R
\name{plot_svd}
\alias{plot_svd}
\title{Singular Value Decomposition}
\usage{
plot_svd(X)
}
\arguments{
\item{X}{A Matrix}
}
\value{
A plot with the first two eigen vectors plotted on the X and Y axis
}
\description{
This function computes the singular value decomposition of a matrix and plots the first two eigen vectors.
}
