% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast_load_write.R
\name{read_proteome}
\alias{read_proteome}
\title{Read the proteome of a given organism}
\usage{
read_proteome(file, format, ...)
}
\arguments{
\item{file}{a character string specifying the path to the file storing the genome.}

\item{format}{a character string specifying the file format used to store the proteome, e.g. "fasta", "fastq".}

\item{...}{additional arguments that are used by the [Biostrings]{readDNAStringSet} function.}
}
\value{
A data.table storing the gene id in the first column and the corresponding sequence as string in the second column.
}
\description{
This function reads an organism specific genome stored in a defined file format.
}
