% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_factorize_matrix.R
\name{row_scale}
\alias{row_scale}
\title{Row normalization}
\usage{
row_scale(z)
}
\arguments{
\item{x}{A Matrix}
}
\value{
The Scaled Matrix
}
\description{
This function loads a file as a matrix. It standardizes by substracting the median row value and dividing by the median absolute deviation.
}
