% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_tools.R
\name{smooth_hist}
\alias{smooth_hist}
\title{Smooth Histogram}
\usage{
smooth_hist(z, unit, bw = "nrd0", n, from, to, plotHist = FALSE,
  add = FALSE, xlab, ylab = "Frequency", xlim, ylim, main)
}
\arguments{
\item{z}{the data}

\item{unit}{the unit which determines the y axis scaling and is drawn}

\item{bw}{arguments to density}

\item{n}{arguments to density}

\item{from}{arguments to density}

\item{to}{arguments to density}

\item{plotHist}{a logical: should an actual histogram be drawn under curve?}

\item{add}{a logical: add should the curve be added to existing plot?}

\item{xlab}{x-axis title, defaults to no title}

\item{ylab}{y-axis title, defaults to no title}

\item{xlim}{range of the x-axis}

\item{ylim}{range of the y-axis}

\item{main}{an overall title for the plot}
}
\value{
Function
}
\description{
a smooth histogram with unit indicator. The advantage is interpretability. If input is a matrix, curves are plotted together.
}
