% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learn_and_predict.R
\name{test_limma_lm}
\alias{test_limma_lm}
\title{Automatic differential gene expression using limma}
\usage{
test_limma_lm(mat, mod, weights)
}
\arguments{
\item{mat}{A gene expression matrix (genes x samples)}

\item{weights}{Assign differential weights either across genes or samples}

\item{model}{A model of contrasts splitting samples}
}
\value{
It outputs a list with the model fit, coefficients per gene, as well as the sum of residual squares after accounting for differential weights
}
\description{
This function tests for differential expression per gene using a linear model and definition contrasts that splits the input matrix into groups of samples. It outputs the fit, coefficients per gene, as well as the sum of residual squares after accounting for weights.
}
