% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learn_and_predict.R
\name{test_rf}
\alias{test_rf}
\title{Learns and classifies using a Random Forest Classifier}
\usage{
test_rf(Xtrain, ytrain, Xtest, ytest)
}
\arguments{
\item{Xtrain}{Predictor variables of training set}

\item{ytrain}{Response variables of training set}

\item{Xtest}{Predictor variables of test set}

\item{ytest}{Response variables of test set}
}
\value{
A list with model fit, test results, AUC, probability per class, original labels
}
\description{
This function trains a Random Forest classifier on the training set and tests its accuracy on a test set. It reports the fit, prediction results, area under curve (AUC accuracy) on test, probability of belonging to predicted class, and original labels of test
}
