% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_factorize_matrix.R
\name{uq_scale}
\alias{uq_scale}
\title{Upper Quartile Normalization}
\usage{
uq_scale(expr_mat, spikes = NULL)
}
\arguments{
\item{expr_mat}{Input matrix with expression values for genes in the form (Genes x Samples)}

\item{spikes}{A vector of genes that were spike-ins in the original experiment}
}
\value{
A matrix with columns upper quartile normalized.
}
\description{
Each column is divided by the 75% quantile of the counts for each library.
The calculated quantile is then scaled by the median across cells to keep the absolute level of expression relatively consistent.
Spike-ins should be excluded from the calculation of total expression in order to correct for total cell RNA content, therefore they should be specified to exclude them from downstream analysis
}
