% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast_load_write.R
\name{write_proteome}
\alias{write_proteome}
\title{Save a proteome in fasta format

This function writes to a file. It assumes that the object to be written #' is a data.table object.}
\usage{
write_proteome(proteome, file.name, nbchar = 80, open = "w",
  as.string = TRUE)
}
\arguments{
\item{proteome}{A data.table object storing amino acid sequences in the first column and gene ids in the second column.}

\item{file.name}{A character string specifying the name of the fasta output file}

\item{as.string}{When set to TRUE sequences are in the form of strings instead of vectors of single characters.}
}
